refactor(tools): separate IP-related functions from tool.utils
KAG-2958
