+++ kong-3.6.0-0.rockspec
@@ -173,6 +173,7 @@ build = {
     ["kong.tools.rand"] = "kong/tools/rand.lua",
     ["kong.tools.system"] = "kong/tools/system.lua",
     ["kong.tools.time"] = "kong/tools/time.lua",
+    ["kong.tools.ip"] = "kong/tools/ip.lua",
 
     ["kong.runloop.handler"] = "kong/runloop/handler.lua",
     ["kong.runloop.events"] = "kong/runloop/events.lua",
